Refactoring Jeweler module
